(PC-7476) bookings: add helper function to get properties labels
	+ add LocationLabel to this function
